+++ src/QueryBuilder/DSL/Aggregation.php
@@ -11,6 +11,7 @@ use Elastica\Aggregation\Composite;
 use Elastica\Aggregation\CumulativeSum;
 use Elastica\Aggregation\DateHistogram;
 use Elastica\Aggregation\DateRange;
+use Elastica\Aggregation\Derivative;
 use Elastica\Aggregation\DiversifiedSampler;
 use Elastica\Aggregation\ExtendedStats;
 use Elastica\Aggregation\Filter;
@@ -35,6 +36,7 @@ use Elastica\Aggregation\ScriptedMetric;
 use Elastica\Aggregation\SerialDiff;
 use Elastica\Aggregation\SignificantTerms;
 use Elastica\Aggregation\Stats;
+use Elastica\Aggregation\StatsBucket;
 use Elastica\Aggregation\Sum;
 use Elastica\Aggregation\SumBucket;
 use Elastica\Aggregation\Terms;
@@ -130,6 +132,14 @@ class Aggregation implements DSL
     public function stats(string $name): Stats
     {
         return new Stats($name);
+    }
+
+    /**
+     * @see https://www.elastic.co/guide/en/elasticsearch/reference/current/search-aggregations-pipeline-stats-bucket-aggregation.html
+     */
+    public function stats_bucket(string $name, ?string $bucketsPath = null): StatsBucket
+    {
+        return new StatsBucket($name, $bucketsPath);
     }
 
     /**
@@ -466,6 +476,14 @@ class Aggregation implements DSL
     public function sampler(string $name): Sampler
     {
         return new Sampler($name);
+    }
+
+    /**
+     * @see https://www.elastic.co/guide/en/elasticsearch/reference/current/search-aggregations-pipeline-derivative-aggregation.html
+     */
+    public function derivative(string $name, ?string $bucketsPath = null): Derivative
+    {
+        return new Derivative($name, $bucketsPath);
     }
 
     /**